:boom: Generify and improve StateMap to IndexBiMap
The data structure now allows fully specifying the type of data for the two maps. But most importantly, it now handles duplicates properly

All tests were passing except for the Lists 'set' tests. The reason is the new data structure using an IdentityHashMap instead of a classic HashMap. Adapted those tests to also check for proper re-usage of cells

Signed-off-by: palexdev <alessandro.parisi406@gmail.com>