:boom: To automatically make use of the new onCreated hook offered by the VFXCell API, we introduced a new class CellFactory that replaces the cell factory function everywhere
This new wrapper class holds the reference to the owning VFXContainer. Cells must be created through the `create(...)` method to ensure the `onCreated(...)` hook is automatically called.
A little exception to this system is VFXColumn which is not a VFXContainer. The CellFactory object is created by inline overriding the `getOwner()` method to return the table instance already offered by the relative property in VFXColumn's. This is because columns are reusable across different tables!

Signed-off-by: palexdev <alessandro.parisi406@gmail.com>