:sparkles: Define base API for all virtualized containers helpers
This is the preliminary change to properly fix the bug patched in commit ac6d596f
The virtual sizes (maxX/maxY) will not be automatically invalidated when the items list changes.
This bug is nasty because it only occurs when someone is listening on those properties (bindings are included ofc) because that would trigger their re-computation early causing in specific occasions to trigger the onPositionChanged(...) method in the behaviors before the onItemsChanged(...) one.
This would lead to a wrong state object being computed with items mapped to the correct item and index but not updated visually.
Under no circumstances a change in the items list should trigger others state computation methods!

Signed-off-by: palexdev <alessandro.parisi406@gmail.com>